ready method before joining

+++ src/main/java/org/example/reminder_my_project/project/CarReminder.java
@@ -17,7 +17,7 @@ public class CarReminder {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idRem;
+    private Long idRem;
 
     @Enumerated(value = EnumType.STRING)
     private ReminderType type;
